% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/coerce.R
\name{as_tibble}
\alias{as_tibble}
\alias{as_tibble,ChromatinContacts-method}
\title{Convert ChromatinContacts to tibble}
\usage{
as_tibble(x, which = "interactions")

\S4method{as_tibble}{ChromatinContacts}(
  x,
  which = c("interactions", "compartments", "TADs", "loops", "multi_contacts")
)
}
\arguments{
\item{x}{ChromatinContacts object.}

\item{which}{Character. Data to convert: \code{"interactions"} (default), or
feature name (\code{"TADs"}, \code{"loops"}, \code{"compartments"}, \code{"multi_contacts"}).}
}
\value{
Tibble with requested data.
}
\description{
Converts interaction or feature data from ChromatinContacts to tibble.
}
\examples{
\dontrun{
cc <- ChromatinContacts("sample.cool") |> import()
df <- as_tibble(cc)
tads_df <- as_tibble(cc, which = "TADs")
}

}
