% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_example_files}
\alias{download_example_files}
\title{download_example_files}
\usage{
download_example_files(check_exists = FALSE)
}
\arguments{
\item{check_exists}{Boolean. Whether to check if the file exists.
Default is \code{FALSE}.}
}
\description{
Download example files under the cache directory.
}
\examples{
\dontrun{
library(gghic)
download_example_files()
}
}
