% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getters-methods.R
\name{features}
\alias{features}
\alias{features,ChromatinContacts,character-method}
\alias{features,ChromatinContacts,missing-method}
\title{Get genomic features}
\usage{
features(x, name)

\S4method{features}{ChromatinContacts,character}(x, name)

\S4method{features}{ChromatinContacts,missing}(x)
}
\arguments{
\item{x}{ChromatinContacts object.}

\item{name}{Character. Feature name: \code{"compartments"}, \code{"TADs"}, \code{"loops"},
or \code{"multi_contacts"}. If missing, returns all as SimpleList.}
}
\value{
GRanges (TADs, compartments, multi_contacts), GInteractions (loops),
or SimpleList (all features).
}
\description{
Retrieves specific or all genomic features.
}
\examples{
\dontrun{
tads <- features(cc, "TADs")
loops <- features(cc, "loops")
all_features <- features(cc)
}

}
