% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters-methods.R
\name{focus,ChromatinContacts-method}
\alias{focus,ChromatinContacts-method}
\title{Get focus regions of ChromatinContacts object}
\usage{
\S4method{focus}{ChromatinContacts}(x)
}
\arguments{
\item{x}{A \code{ChromatinContacts} object.}
}
\value{
GInteractions object representing the focus regions, or NULL if
genome-wide.
}
\description{
Retrieves the focus regions (genomic regions of interest) from a
\code{ChromatinContacts} object.
}
\examples{
\dontrun{
cc <- ChromatinContacts("sample.cool", focus = "chr1")
focus(cc)
}

}
