% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getters-methods.R
\name{focus}
\alias{focus}
\alias{focus,ChromatinContacts-method}
\title{Get focus regions}
\usage{
focus(x)

\S4method{focus}{ChromatinContacts}(x)
}
\arguments{
\item{x}{ChromatinContacts object.}
}
\value{
GInteractions (focus regions) or NULL (genome-wide).
}
\description{
Retrieves focus regions from ChromatinContacts object.
}
\examples{
\dontrun{
cc <- ChromatinContacts("sample.cool", focus = "chr1")
focus(cc)
}

}
