% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{generics}
\alias{generics}
\alias{resolution}
\alias{focus}
\alias{features}
\alias{features<-}
\alias{as_tibble}
\title{Generic functions for ChromatinContacts objects}
\usage{
resolution(x)

focus(x)

features(x, name)

features(x, name, ...) <- value

as_tibble(x, which = "interactions")
}
\description{
Generic functions to access and manipulate ChromatinContacts objects.
}
\keyword{internal}
