% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-annotation.R
\name{geom_annotation}
\alias{geom_annotation}
\title{geom_annotation}
\usage{
geom_annotation(
  mapping = NULL,
  data = NULL,
  stat = StatAnnotation,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  check.param = FALSE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{ggplot2::aes()}}.}

\item{...}{Parameters including:
\itemize{
\item \code{txdb}: The TxDb object. Default is \code{NULL}.
\item tx2gene: An optional data frame or tibble that maps transcript information to gene information. It should include the following columns:
\itemize{
\item chrom: Chromosome number or name.
\item gene_id: Entrez gene ID.
\item gene_symbol: Common symbol or name of the gene.
\item tx_id: Entrez transcript ID.
\item tx_name: Name of the transcript.
\item gene_type: Type or classification of the gene.
\item tx_type: Type or classification of the transcript.
}
\item \code{gtf_path}: The path to the GTF file, which is used to generate \code{txdb} and \code{tx2gene}. Generated files are saved in the cache directory. Default is \code{NULL}.
\item \code{width_ratio}: The ratio of the width of each gene model track relative to the height of the Hi-C plot. Default is \code{0.02}.
\item \code{maxgap}: The maximum gap between genes to be drawn in the same line. Default is \code{-1}.
\item \code{include_ncrna}: Whether to include ncRNA or not. Default is \code{TRUE}.
\item \code{style}: The style of the gene model track, which can be \code{"basic"} or \code{"arrow"}. Default is \code{"basic"}.
}}
}
\value{
A ggplot object.
}
\description{
A ggplot2 geom for gene model tracks.
}
\details{
Requires the following aesthetics:
\itemize{
\item seqnames1
\item start1
\item end1
\item seqnames2
\item start2
\item end2
}
}
\examples{
\dontrun{
}
}
