% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-hic.R
\name{geom_hic}
\alias{geom_hic}
\title{geom_hic}
\usage{
geom_hic(
  mapping = NULL,
  data = NULL,
  stat = StatHic,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{ggplot2::aes()}}.}
}
\value{
A ggplot object.
}
\description{
A ggplot2 geom for Hi-C data.
}
\details{
Requires the following aesthetics:
\itemize{
\item seqnames1
\item start1
\item end1
\item seqnames2
\item start2
\item end2
\item fill
}
}
\examples{
\dontrun{
}
}
