% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-ideogram.R
\name{geom_ideogram}
\alias{geom_ideogram}
\title{geom_ideogram}
\usage{
geom_ideogram(
  mapping = NULL,
  data = NULL,
  stat = StatIdeogram,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{ggplot2::aes()}}.}

\item{...}{Parameters including:
\itemize{
\item \code{genome}: The genome name. Default is \code{"hg19"}.
\item \code{highlight}: Whether to highlight the boundary of the chromosome. Default is \code{TRUE}.
\item \code{fontsize}: The font size of the chromosome name. Default is \code{10}.
}}
}
\value{
A ggplot object.
}
\description{
A ggplot2 geom for chromosome ideogram.
}
\details{
Requires the following aesthetics:
\itemize{
\item seqnames1
\item start1
\item end1
\item seqnames2
\item start2
\item end2
}
}
\examples{
\dontrun{
}
}
