% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-tad2.R
\name{geom_tad2}
\alias{geom_tad2}
\title{Visualize TAD boundaries (a second version)}
\usage{
geom_tad2(
  mapping = NULL,
  data = NULL,
  stat = StatTad2,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used to override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}

\item{...}{Parameters to be ignored.}
}
\value{
A ggplot object.
}
\description{
Another version of \code{\link[=geom_tad]{geom_tad()}} that requires direct data frame input.
Draws TAD boundaries as triangular outlines on the Hi-C heatmap.
}
\details{
Requires the following aesthetics:
\itemize{
\item seqname
\item start
\item end
}
}
\examples{
\dontrun{
# Load Hi-C data
cc <- ChromatinContacts("path/to/cooler.cool", focus = "chr4") |>
  import()

# Load TAD data
tad_file <- "path/to/tads.bed"
tads <- read.table(tad_file, col.names = c("seqnames", "start", "end"))

# Add TAD boundaries using data frame
library(ggplot2)
gghic(cc) +
  geom_tad2(data = tads, aes(seqnames = seqnames, start = start, end = end))

# Multiple TAD types with different colors
tads1 <- tads[1:5, ]
tads1$type <- "Strong"
tads2 <- tads[6:10, ]
tads2$type <- "Weak"
all_tads <- rbind(tads1, tads2)

gghic(cc) +
  geom_tad2(
    data = all_tads,
    aes(seqnames = seqnames, start = start, end = end, colour = type)
  ) +
  scale_colour_manual(values = c("Strong" = "red", "Weak" = "grey"))
}
}
