% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_breaks_labels}
\alias{get_breaks_labels}
\title{get_breaks_labels}
\usage{
get_breaks_labels(data)
}
\arguments{
\item{data}{A data frame/tibble with columns \code{seqnames} and \code{end}.}
}
\value{
A list with \code{breaks} and \code{labels}.
}
\description{
Get breaks and labels for axis from multiple chromosomes.
}
