% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghic-package.R
\docType{package}
\name{gghic-package}
\alias{gghic-package}
\title{gghic: gghic}
\description{
gghic
}
\author{
\strong{Maintainer}: Minghao Jiang \email{mjhk@connect.hku.hk} (\href{https://orcid.org/0000-0003-3309-9185}{ORCID})

}
\keyword{internal}
