% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghic.R
\name{gghic}
\alias{gghic}
\title{gghic}
\usage{
gghic(
  data = NULL,
  score_column = "balanced",
  scale_method = log10,
  ideogram = TRUE,
  annotation = TRUE,
  ...
)
}
\arguments{
\item{data}{A HiCExperiment object or a tibble/data.frame}

\item{score_column}{The column name of which the score is calculated.
Default is \code{"balanced"}.}

\item{scale_method}{The function to scale the score. Default is \code{log10}.}

\item{ideogram}{Whether to add ideogram or not. Default is \code{TRUE}.}

\item{annotation}{Whether to add annotation or not. Default is \code{TRUE}.}

\item{...}{Other parameters passed to \code{\link[=geom_ideogram]{geom_ideogram()}}, \code{\link[=geom_annotation]{geom_annotation()}}.}
}
\value{
A ggplot object.
}
\description{
A ggplot2 wrapper for HiC data.
}
\examples{
\dontrun{
library(HiCExperiment)
}
}
