% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergraph.R
\name{print.hypergraph}
\alias{print.hypergraph}
\title{Print method for hypergraph objects}
\usage{
\method{print}{hypergraph}(x, ...)
}
\arguments{
\item{x}{A hypergraph object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for hypergraph objects
}
