% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters-methods.R
\name{resolution,ChromatinContacts-method}
\alias{resolution,ChromatinContacts-method}
\title{Get resolution of ChromatinContacts object}
\usage{
\S4method{resolution}{ChromatinContacts}(x)
}
\arguments{
\item{x}{A \code{ChromatinContacts} object.}
}
\value{
Integer representing resolution in base pairs, or NULL if not set.
}
\description{
Retrieves the resolution (bin size) in base pairs from a \code{ChromatinContacts}
object.
}
\examples{
\dontrun{
cc <- ChromatinContacts("sample.mcool", resolution = 10000)
resolution(cc) # Returns 10000
}

}
