% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getters-methods.R
\name{resolution}
\alias{resolution}
\alias{resolution,ChromatinContacts-method}
\title{Get resolution of ChromatinContacts object}
\usage{
resolution(x)

\S4method{resolution}{ChromatinContacts}(x)
}
\arguments{
\item{x}{ChromatinContacts object.}
}
\value{
Integer (resolution in bp) or NULL.
}
\description{
Retrieves bin size in base pairs.
}
\examples{
\dontrun{
cc <- ChromatinContacts("sample.mcool", resolution = 10000L)
resolution(cc) # Returns 10000
}

}
