% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters-methods.R
\name{seqinfo,ChromatinContacts-method}
\alias{seqinfo,ChromatinContacts-method}
\title{Get sequence information from ChromatinContacts object}
\usage{
\S4method{seqinfo}{ChromatinContacts}(x)
}
\arguments{
\item{x}{A \code{ChromatinContacts} object.}
}
\value{
A \code{Seqinfo} object containing sequence/chromosome information.
}
\description{
Retrieves the sequence information (chromosome names, lengths) from a
\code{ChromatinContacts} object.
}
\examples{
\dontrun{
cc <- ChromatinContacts("sample.cool")
seqinfo(cc)
}

}
