% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-hic.R
\name{theme_hic}
\alias{theme_hic}
\title{theme_hic}
\usage{
theme_hic(
  hide_y = TRUE,
  coord_ratio = 1,
  scale_fill_gradientn = TRUE,
  breaks = ggplot2::waiver(),
  labels = scales::unit_format(unit = "M", scale = 1e-06),
  xmin = NULL,
  xmax = NULL,
  expand_x = c(0, 0)
)
}
\arguments{
\item{hide_y}{A logical value indicating whether to hide the y-axis.
Default is \code{TRUE}.}

\item{coord_ratio}{The ratio of the \link[ggplot2:coord_fixed]{ggplot2::coord_fixed}. Default is \code{1}.}

\item{scale_fill_gradientn}{A logical value indicating whether to apply the
gradient color scale. Default is \code{TRUE}.}

\item{breaks}{This argument is passed to \link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous}.}

\item{labels}{This argument is passed to \link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous}.}

\item{xmin}{The minimum x-axis value. Default is \code{NULL}.}

\item{xmax}{The maximum x-axis value. Default is \code{NULL}.}

\item{expand_x}{A numeric vector of length 2 to expand the x-axis.
Default is \code{c(0, 0)}.}
}
\value{
A ggplot2 theme.
}
\description{
Generate a ggplot2 theme for Hi-C plots.
}
\details{
If either \code{xmin} or \code{xmax} is \code{NULL}, the x-axis will not be expanded.
}
\examples{
\dontrun{
# Load Hi-C data
cc <- ChromatinContacts("path/to/cooler.cool", focus = "chr4") |>
  import()

# Apply default theme_hic
gghic(cc) + theme_hic()

# Disable default color scale
gghic(cc) +
  theme_hic(scale_fill_gradientn = FALSE) +
  scale_fill_viridis_c()
}
}
