% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatemers.R
\name{concatemers2gis}
\alias{concatemers2gis}
\title{concatemers2gis}
\usage{
concatemers2gis(grs, region = NULL, bin_size = 1e+05, read_group = "read_name")
}
\arguments{
\item{grs}{A GRanges object of concatemer reads.}

\item{region}{A GRanges object specifying the genomic region of interest.
Default is \code{NULL}, which uses the entire range of \code{grs}.}

\item{bin_size}{The size of the bins to divide the genomic region into.
Default is \code{100000} (100 kb).}

\item{read_group}{The metadata column in \code{grs} that defines the read groups.
Default is \code{"read_name"}.}
}
\value{
A \link[InteractionSet:GInteractions-class]{InteractionSet::GInteractions} object.
}
\description{
Convert concatemer reads to pairwise interactions in a
GInteractions object.
}
\examples{
\dontrun{
}
}
