% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_example_files}
\alias{download_example_files}
\title{download_example_files}
\usage{
download_example_files()
}
\description{
Download example files under the cache directory.
}
\examples{
\dontrun{
library(gghic)
download_example_files()
}
}
