% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters-methods.R
\name{features,ChromatinContacts,missing-method}
\alias{features,ChromatinContacts,missing-method}
\title{Get all genomic features from ChromatinContacts object}
\usage{
\S4method{features}{ChromatinContacts,missing}(x)
}
\arguments{
\item{x}{A \code{ChromatinContacts} object.}
}
\value{
A \code{SimpleList} containing all features: compartments, TADs, loops,
and multi_contacts.
}
\description{
Retrieves all genomic features (TADs, loops, compartments, multi-contacts)
as a SimpleList.
}
\examples{
\dontrun{
# Get all features
all_features <- features(cc)
all_features$TADs
all_features$loops
}

}
