% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-annotation.R
\name{geom_annotation}
\alias{geom_annotation}
\title{geom_annotation}
\usage{
geom_annotation(
  mapping = NULL,
  data = NULL,
  stat = StatAnnotation,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...,
  txdb = NULL,
  tx2gene = NULL,
  gtf_path = NULL,
  width_ratio = 1/50,
  spacing_ratio = 1/3,
  maxgap = -1,
  include_ncrna = TRUE,
  style = c("basic", "arrow"),
  gene_symbols = NULL,
  chrom_prefix = TRUE,
  fontsize = 10,
  colour = "#48CFCB",
  fill = "#48CFCB",
  draw_boundary = TRUE,
  boundary_colour = "black",
  linetype = "dashed"
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used the override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{Parameters to be ignored.}

\item{txdb}{The TxDb object. Default is \code{NULL}.}

\item{tx2gene}{An optional data frame or tibble that maps transcript
information to gene information. It should include the following columns:
\itemize{
\item chrom: Chromosome number or name.
\item gene_id: Entrez gene ID.
\item gene_symbol: Common symbol or name of the gene.
\item tx_id: Entrez transcript ID.
\item tx_name: Name of the transcript.
\item gene_type: Type or classification of the gene.
\item tx_type: Type or classification of the transcript.
}}

\item{gtf_path}{The path to the GTF file, which is used to generate \code{txdb}
and \code{tx2gene}. Generated files are saved in the cache directory.
Default is \code{NULL}.}

\item{width_ratio}{The ratio of the width of each gene model track
relative to the height of the Hi-C plot. Default is \code{1/50}.}

\item{spacing_ratio}{The ratio of the spacing between two gene model tracks.
Default is \code{1/3}.}

\item{maxgap}{The maximum gap between genes to be drawn in the same line.
Default is \code{-1}.}

\item{include_ncrna}{Whether to include ncRNA or not. Default is \code{TRUE}.}

\item{style}{The style of the gene model track, which can be \code{"basic"}
or \code{"arrow"}. Default is \code{"basic"}.}

\item{gene_symbols}{A character vector of gene symbols to be included only.
Default is \code{NULL}.}

\item{chrom_prefix}{Whether the input data has chromosome names
with prefix 'chr' or not. Default is \code{TRUE}.}

\item{fontsize}{The font size of the gene symbols. Default is \code{10}.}

\item{colour}{The color of the gene model track. Default is \code{"#48CFCB"}.}

\item{fill}{The fill color of the gene model track. Default is \code{"#48CFCB"}.}

\item{draw_boundary}{Whether to draw the boundary line or not when plotting
multiple chromosomes. Default is \code{TRUE}.}

\item{boundary_colour}{The color of the boundary line. Default is \code{"black"}.}

\item{linetype}{The line type of the boundary line. Default is \code{"dashed"}.}
}
\value{
A ggplot object.
}
\description{
A ggplot2 geom for gene model tracks.
}
\details{
Requires the following aesthetics:
\itemize{
\item seqnames1
\item start1
\item end1
\item seqnames2
\item start2
\item end2
}
}
\examples{
\dontrun{
library(gghic)
library(ggplot2)
library(dplyr)
library(HiCExperiment)
library(InteractionSet)
library(scales)
library(glue)
library(rappdirs)

dir_cache <- user_cache_dir(appname = "gghic")
url_file <- paste0(
  "https://raw.githubusercontent.com/mhjiang97/gghic-data/refs/heads/",
  "master/cooler/chr4_11-5kb.cool"
)
path_file <- glue("{dir_cache}/chr4_11-5kb.cool")
download.file(url_file, path_file)

hic <- path_file |>
  CoolFile() |>
  import()

gis <- interactions(hic)
gis$score <- log10(gis$balanced)
x <- as_tibble(gis)
scores <- x$score[pairdist(gis) != 0 & !is.na(pairdist(gis) != 0)]
scores <- scores[!is.na(scores) & !is.infinite(scores)]
x$score <- oob_squish(x$score, c(min(scores), max(scores)))

p <- x |>
  filter(
    seqnames1 == "chr11", seqnames2 == "chr11",
    center1 > 67000000, center1 < 67100000,
    center2 > 67000000, center2 < 67100000
  ) |>
  ggplot(
    aes(
      seqnames1 = seqnames1, start1 = start1, end1 = end1,
      seqnames2 = seqnames2, start2 = start2, end2 = end2,
      fill = score
    )
  ) +
  geom_hic()

path_gtf <- glue("{dir_cache_gghic}/gencode-chr4_11.gtf.gz")

p + geom_annotation(gtf_path = path_gtf, style = "basic", maxgap = 100000)
}
}
