% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-annotation.R
\name{geom_annotation}
\alias{geom_annotation}
\title{geom_annotation}
\usage{
geom_annotation(
  mapping = NULL,
  data = NULL,
  stat = StatAnnotation,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  check.param = FALSE,
  ...,
  txdb = NULL,
  tx2gene = NULL,
  gtf_path = NULL,
  width_ratio = 1/50,
  spacing_ratio = 1/3,
  maxgap = -1,
  include_ncrna = TRUE,
  style = c("basic", "arrow"),
  gene_symbols = NULL,
  fontsize = 10,
  colour = "#48CFCB",
  fill = "#48CFCB",
  draw_boundary = TRUE,
  boundary_colour = "black",
  linetype = "dashed"
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{ggplot2::aes()}}.}

\item{...}{Parameters to be ignored.}

\item{txdb}{The TxDb object. Default is \code{NULL}.}

\item{tx2gene}{An optional data frame or tibble that maps transcript
information to gene information. It should include the following columns:
\itemize{
\item chrom: Chromosome number or name.
\item gene_id: Entrez gene ID.
\item gene_symbol: Common symbol or name of the gene.
\item tx_id: Entrez transcript ID.
\item tx_name: Name of the transcript.
\item gene_type: Type or classification of the gene.
\item tx_type: Type or classification of the transcript.
}}

\item{gtf_path}{The path to the GTF file, which is used to generate \code{txdb}
and \code{tx2gene}. Generated files are saved in the cache directory.
Default is \code{NULL}.}

\item{width_ratio}{The ratio of the width of each gene model track
relative to the height of the Hi-C plot. Default is \code{1/50}.}

\item{spacing_ratio}{The ratio of the spacing between two gene model tracks.
Default is \code{1/3}.}

\item{maxgap}{The maximum gap between genes to be drawn in the same line.
Default is \code{-1}.}

\item{include_ncrna}{Whether to include ncRNA or not. Default is \code{TRUE}.}

\item{style}{The style of the gene model track, which can be \code{"basic"}
or \code{"arrow"}. Default is \code{"basic"}.}

\item{gene_symbols}{A character vector of gene symbols to be included only.
Default is \code{NULL}.}

\item{fontsize}{The font size of the gene symbols. Default is \code{10}.}

\item{colour}{The color of the gene model track. Default is \code{"#48CFCB"}.}

\item{fill}{The fill color of the gene model track. Default is \code{"#48CFCB"}.}

\item{draw_boundary}{Whether to draw the boundary line or not when plotting
multiple chromosomes. Default is \code{TRUE}.}

\item{boundary_colour}{The color of the boundary line. Default is \code{"black"}.}

\item{linetype}{The line type of the boundary line. Default is \code{"dashed"}.}
}
\value{
A ggplot object.
}
\description{
A ggplot2 geom for gene model tracks.
}
\details{
Requires the following aesthetics:
\itemize{
\item seqnames1
\item start1
\item end1
\item seqnames2
\item start2
\item end2
}
}
\examples{
\dontrun{
library(gghic)
library(ggplot2)
library(dplyr)
library(HiCExperiment)
library(InteractionSet)
library(scales)
library(scales)

cf <- HiCExperiment::CoolFile(system.file("extdata", "cooler", "chr4_11-5kb.cool", package = "gghic"))
hic <- HiCExperiment::import(cf)

gis <- InteractionSet::interactions(hic)
gis$score <- log10(gis$balanced)
x <- tibble::as_tibble(gis)
scores <- x$score[InteractionSet::pairdist(gis) != 0 & !is.na(InteractionSet::pairdist(gis) != 0)]
scores <- scores[!is.na(scores) & !is.infinite(scores)]
x$score <- scales::oob_squish(x$score, c(min(scores), max(scores)))

p <- x |>
  dplyr::filter(
    seqnames1 == "chr11", seqnames2 == "chr11",
    center1 > 67000000, center1 < 67100000,
    center2 > 67000000, center2 < 67100000
  ) |>
  ggplot2::ggplot(
    ggplot2::aes(
      seqnames1 = seqnames1, start1 = start1, end1 = end1,
      seqnames2 = seqnames2, start2 = start2, end2 = end2,
      fill = score
    )
  ) +
  geom_hic()

path_gtf <- system.file(
  "extdata", "gtf", "gencode-chr4_11.gtf.gz", package = "gghic"
)
p + geom_annotation(gtf_path = path_gtf, style = "basic", maxgap = 100000)
}
}
