% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-hic.R
\name{geom_hic}
\alias{geom_hic}
\title{geom_hic}
\usage{
geom_hic(
  mapping = NULL,
  data = NULL,
  stat = StatHic,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...,
  draw_boundary = TRUE,
  boundary_colour = "black",
  linetype = "dashed"
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{ggplot2::aes()}}.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used the override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{Parameters to be ignored.}

\item{draw_boundary}{Whether to draw the boundary line or not when plotting
multiple chromosomes. Default is \code{TRUE}.}

\item{boundary_colour}{The color of the boundary line. Default is \code{"black"}.}

\item{linetype}{The line type of the boundary line. Default is \code{"dashed"}.}
}
\value{
A ggplot object.
}
\description{
A ggplot2 geom for Hi-C data.
}
\details{
Requires the following aesthetics:
\itemize{
\item seqnames1
\item start1
\item end1
\item seqnames2
\item start2
\item end2
\item fill
}
}
\examples{
\dontrun{
library(gghic)
library(ggplot2)
library(dplyr)
library(HiCExperiment)
library(InteractionSet)
library(scales)
library(scales)

cf <- HiCExperiment::CoolFile(
  system.file("extdata", "cooler", "chr4_11-5kb.cool", package = "gghic")
)
hic <- HiCExperiment::import(cf)

gis <- InteractionSet::interactions(hic)
gis$score <- log10(gis$balanced)
x <- tibble::as_tibble(gis)
scores <- x$score[
  InteractionSet::pairdist(gis) != 0 &
    !is.na(InteractionSet::pairdist(gis) != 0)
]
scores <- scores[!is.na(scores) & !is.infinite(scores)]
x$score <- scales::oob_squish(x$score, c(min(scores), max(scores)))

x |>
  dplyr::filter(
    seqnames1 == "chr11", seqnames2 == "chr11",
    center1 > 67000000, center1 < 67100000,
    center2 > 67000000, center2 < 67100000
  ) |>
  ggplot2::ggplot(
    ggplot2::aes(
      seqnames1 = seqnames1, start1 = start1, end1 = end1,
      seqnames2 = seqnames2, start2 = start2, end2 = end2,
      fill = score
    )
  ) +
  geom_hic() +
  theme_hic()
}
}
