% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-ideogram.R
\name{geom_ideogram}
\alias{geom_ideogram}
\title{geom_ideogram}
\usage{
geom_ideogram(
  mapping = NULL,
  data = NULL,
  stat = StatIdeogram,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...,
  genome = "hg19",
  highlight = FALSE,
  width_ratio = 1/30,
  length_ratio = 0.8,
  fontsize = 10,
  colour = "red",
  fill = "#FFE3E680"
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used the override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{Parameters to be ignored.}

\item{genome}{The genome name. Default is \code{"hg19"}.}

\item{highlight}{Whether to highlight the boundary of the chromosome.
Default is \code{TRUE}.}

\item{width_ratio}{The ratio of the width of each chromosome ideogram
relative to the height of the Hi-C plot. Default is \code{1/30}.}

\item{length_ratio}{The ratio of the length of each chromosome ideogram
relative to the width of the Hi-C plot. Default is \code{0.8}.}

\item{fontsize}{The font size of the chromosome name. Default is \code{10}.}

\item{colour}{The color of the chromosome boundary. Default is \code{"red"}.}

\item{fill}{The fill color of the highlighted region on the ideogram.
Default is \code{"#FFE3E680"}.}
}
\value{
A ggplot object.
}
\description{
A ggplot2 geom for chromosome ideogram.
}
\details{
Requires the following aesthetics:
\itemize{
\item seqnames1
\item start1
\item end1
\item seqnames2
\item start2
\item end2
}
}
\examples{
\dontrun{
library(gghic)
library(ggplot2)
library(dplyr)
library(HiCExperiment)
library(InteractionSet)
library(scales)
library(glue)
library(rappdirs)

download_example_files()
dir_cache_gghic <- user_cache_dir(appname = "gghic")

hic <- glue("{dir_cache_gghic}/chr4_11-100kb.cool") |>
  CoolFile() |>
  import(cf)

gis <- interactions(hic)
gis$score <- log10(gis$balanced)
x <- as_tibble(gis)
scores <- x$score[pairdist(gis) != 0 & !is.na(pairdist(gis) != 0)]
scores <- scores[!is.na(scores) & !is.infinite(scores)]
x$score <- oob_squish(x$score, c(min(scores), max(scores)))

p <- x |>
  filter(seqnames1 == "chr11", seqnames2 == "chr11") |>
  ggplot(
    aes(
      seqnames1 = seqnames1, start1 = start1, end1 = end1,
      seqnames2 = seqnames2, start2 = start2, end2 = end2,
      fill = score
    )
  ) +
  geom_hic() +
  theme_hic()

p + geom_ideogram(
  genome = "hg19", highlight = FALSE, length_ratio = 0.7, fontsize = 8
)
}
}
