% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-ideogram.R
\name{geom_ideogram}
\alias{geom_ideogram}
\title{geom_ideogram}
\usage{
geom_ideogram(
  mapping = NULL,
  data = NULL,
  stat = StatIdeogram,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  check.param = FALSE,
  ...,
  genome = "hg19",
  highlight = FALSE,
  width_ratio = 1/30,
  length_ratio = 0.8,
  fontsize = 10,
  colour = "red",
  fill = "#FFE3E680"
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{ggplot2::aes()}}.}

\item{...}{Parameters to be ignored.}

\item{genome}{The genome name. Default is \code{"hg19"}.}

\item{highlight}{Whether to highlight the boundary of the chromosome.
Default is \code{TRUE}.}

\item{width_ratio}{The ratio of the width of each chromosome ideogram
relative to the height of the Hi-C plot. Default is \code{1/30}.}

\item{length_ratio}{The ratio of the length of each chromosome ideogram
relative to the width of the Hi-C plot. Default is \code{0.8}.}

\item{fontsize}{The font size of the chromosome name. Default is \code{10}.}

\item{colour}{The color of the chromosome boundary. Default is \code{"red"}.}

\item{fill}{The fill color of the highlighted region on the ideogram.
Default is \code{"#FFE3E680"}.}
}
\value{
A ggplot object.
}
\description{
A ggplot2 geom for chromosome ideogram.
}
\details{
Requires the following aesthetics:
\itemize{
\item seqnames1
\item start1
\item end1
\item seqnames2
\item start2
\item end2
}
}
\examples{
\dontrun{
library(gghic)
library(ggplot2)
library(dplyr)
library(HiCExperiment)
library(InteractionSet)
library(scales)
library(scales)

cf <- HiCExperiment::CoolFile(system.file("extdata", "cooler", "chr4_11-100kb.cool", package = "gghic"))
hic <- HiCExperiment::import(cf)

gis <- InteractionSet::interactions(hic)
gis$score <- log10(gis$balanced)
x <- tibble::as_tibble(gis)
scores <- x$score[InteractionSet::pairdist(gis) != 0 & !is.na(InteractionSet::pairdist(gis) != 0)]
scores <- scores[!is.na(scores) & !is.infinite(scores)]
x$score <- scales::oob_squish(x$score, c(min(scores), max(scores)))

p <- x |>
  dplyr::filter(seqnames1 == "chr11", seqnames2 == "chr11") |>
  ggplot2::ggplot(
    ggplot2::aes(
      seqnames1 = seqnames1, start1 = start1, end1 = end1,
      seqnames2 = seqnames2, start2 = start2, end2 = end2,
      fill = score
    )
  ) +
  geom_hic() +
  theme_hic()

p + geom_ideogram(
  genome = "hg19", highlight = FALSE, length_ratio = 0.7, fontsize = 8
)
}
}
