% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-track.R
\name{geom_track}
\alias{geom_track}
\title{geom_track}
\usage{
geom_track(
  mapping = NULL,
  data = NULL,
  stat = StatTrack,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  check.param = FALSE,
  ...,
  data_paths = NULL,
  data_granges = NULL,
  width_ratio = 1/20,
  spacing_ratio = 0.5,
  data_range = c("auto", "maximum"),
  fill = "black",
  fontsize = 5,
  draw_boundary = TRUE,
  boundary_colour = "black",
  linetype = "dashed"
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{ggplot2::aes()}}.}

\item{...}{Parameters to be ignored.}

\item{data_paths}{The paths to the sequencing data files. Default is \code{NULL}.}

\item{data_granges}{The GRanges object of the sequencing data.
Default is \code{NULL}.}

\item{width_ratio}{The ratio of the width of each track relative to the
height of the Hi-C plot. Default is \code{1/20}.}

\item{spacing_ratio}{The ratio of the spacing between two tracks.
Default is \code{0.5}.}

\item{data_range}{The range of the x axis. It can be \code{"auto"}, \code{"maximum"},
or a number (vector). Default is \code{"auto"}.}

\item{fill}{The fill color of the track. Default is \code{"black"}.}

\item{fontsize}{The font size of the track names. Default is \code{5}.}

\item{draw_boundary}{Whether to draw the boundary line or not when plotting
multiple chromosomes. Default is \code{TRUE}.}

\item{boundary_colour}{The color of the boundary line. Default is \code{"black"}.}

\item{linetype}{The line type of the boundary line. Default is \code{"dashed"}.}
}
\value{
A ggplot object.
}
\description{
A ggplot2 geom for sequencing data tracks.
}
\details{
Requires the following aesthetics:
\itemize{
\item seqnames1
\item start1
\item end1
\item seqnames2
\item start2
\item end2
}
}
\examples{
\dontrun{
library(gghic)
library(ggplot2)
library(dplyr)
library(HiCExperiment)
library(InteractionSet)
library(scales)
library(scales)

cf <- HiCExperiment::CoolFile(
  system.file("extdata", "cooler", "chr4_11-5kb.cool", package = "gghic")
)
hic <- HiCExperiment::import(cf)

gis <- InteractionSet::interactions(hic)
gis$score <- log10(gis$balanced)
x <- tibble::as_tibble(gis)
scores <- x$score[
  InteractionSet::pairdist(gis) != 0 &
    !is.na(InteractionSet::pairdist(gis) != 0)
]
scores <- scores[!is.na(scores) & !is.infinite(scores)]
x$score <- scales::oob_squish(x$score, c(min(scores), max(scores)))

p <- x_5 |>
  dplyr::filter(
    center1 > 10000000 & center1 < 11000000 &
      center2 > 10000000 & center2 < 11000000
  ) |>
  gghic(
    draw_boundary = TRUE,

    ideogram = TRUE, genome = "hg19", highlight = TRUE,
    ideogram_fontsize = 7, ideogram_width_ratio = 0.08,

    annotation = TRUE, include_ncrna = FALSE, gtf_path = path_gtf,
    style = "arrow", maxgap = 100000, annotation_fontsize = 5,
    annotation_width_ratio = 0.05,

    expand_xaxis = TRUE
  )

paths_track <- system.file(
  "extdata", "bigwig", c("track1.bigWig", "track2.bigWig"),
  package = "gghic"
)

p + geom_track(
  data_paths = paths_track, width_ratio = 0.5,
  fill = c("#DC0000B2", "#00A087B2"), data_range = "auto"
)
}
}
