% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_track2.R
\name{geom_track2}
\alias{geom_track2}
\title{geom_track2}
\usage{
geom_track2(
  mapping = NULL,
  data = NULL,
  stat = StatTrack2,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...,
  width_ratio = 1/20,
  spacing_ratio = 0.5,
  data_range = c("auto", "maximum"),
  rasterize = TRUE,
  dpi = 300,
  dev = "cairo",
  scale = 1,
  draw_boundary = TRUE,
  boundary_colour = "black",
  linetype = "dashed"
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used the override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{Parameters to be ignored.}

\item{width_ratio}{The ratio of the width of each track relative to the
height of the Hi-C plot. Default is \code{1/20}.}

\item{spacing_ratio}{The ratio of the spacing between two tracks.
Default is \code{0.5}.}

\item{data_range}{The range of the x axis. It can be \code{"auto"}, \code{"maximum"},
or a number (vector). Default is \code{"auto"}.}

\item{rasterize}{Whether to rasterize the plot or not. Default is \code{TRUE}.}

\item{dpi}{The resolution of the rasterised plot. Default is \code{300}.}

\item{dev}{The device to rasterise the plot. Default is \code{"cairo"}.}

\item{scale}{The scale of the rasterised plot. Default is \code{1}.}

\item{draw_boundary}{Whether to draw the boundary line or not when plotting
multiple chromosomes. Default is \code{TRUE}.}

\item{boundary_colour}{The color of the boundary line. Default is \code{"black"}.}

\item{linetype}{The line type of the boundary line. Default is \code{"dashed"}.}
}
\value{
A ggplot object.
}
\description{
A second version of \code{\link[=geom_track]{geom_track()}}.
}
\details{
Requires the following aesthetics:
\itemize{
\item seqnames
\item start
\item end
\item score
\item name
}
}
\examples{
\dontrun{
}
}
