% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghic-package.R
\docType{package}
\name{gghic-package}
\alias{gghic-package}
\title{gghic: GGHIC: A Package for Visualizing Genomic Interaction Data Using ggplot2}
\description{
An easy way to create publication-ready figures for genomic interaction data using ggplot2.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jasonwong-lab/gghic}
  \item Report bugs at \url{https://github.com/jasonwong-lab/gghic/issues}
}

}
\author{
\strong{Maintainer}: Minghao Jiang \email{mjhk@connect.hku.hk} (\href{https://orcid.org/0000-0003-3309-9185}{ORCID}) [contributor, copyright holder]

}
\keyword{internal}
