% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters-methods.R
\name{interactions,ChromatinContacts-method}
\alias{interactions,ChromatinContacts-method}
\title{Get interaction data}
\usage{
\S4method{interactions}{ChromatinContacts}(x)
}
\arguments{
\item{x}{ChromatinContacts object.}
}
\value{
GInteractions (Hi-C data) or NULL (not yet imported).
}
\description{
Retrieves imported Hi-C interactions.
}
\details{
Must load with \code{import()} first. GInteractions contains bin coordinates and
metadata (count, balanced).
}
\examples{
\dontrun{
cc <- ChromatinContacts("sample.cool") |> import()
gis <- interactions(cc)
}

}
