% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters-methods.R
\name{seqinfo,ChromatinContacts-method}
\alias{seqinfo,ChromatinContacts-method}
\title{Get sequence information}
\usage{
\S4method{seqinfo}{ChromatinContacts}(x)
}
\arguments{
\item{x}{ChromatinContacts object.}
}
\value{
Seqinfo object.
}
\description{
Retrieves chromosome/sequence information.
}
\examples{
\dontrun{
cc <- ChromatinContacts("sample.cool")
seqinfo(cc)
}

}
